--- docs/source/contrib.glmm.rst
@@ -1,8 +1,0 @@
-Generalised Linear Mixed Models
-===============================
-
-.. automodule:: pyro.contrib.glmm
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :member-order: bysource
