+++ docs/source/contrib.oed.rst
@@ -8,3 +8,11 @@ Expected Information Gain
 .. automodule:: pyro.contrib.oed.eig
     :members:
     :member-order: bysource
+
+Generalised Linear Mixed Models
+-------------------------------
+.. automodule:: pyro.contrib.oed.glmm
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
