+++ pyro/contrib/oed/glmm/__init__.py
@@ -1,11 +1,13 @@
 """
-The :mod:`pyro.contrib.glmm` module provides models and guides for
+.. warning :: This module will eventually be deprecated and removed in favor of [brmp](https://github.com/pyro-ppl/brmp)
+
+The :mod:`pyro.contrib.oed.glmm` module provides models and guides for
 generalised linear mixed models (GLMM). It also includes the
 Normal-inverse-gamma family.
 
 To create a classical Bayesian linear model, use::
 
-    from pyro.contrib.glmm import known_covariance_linear_model
+    from pyro.contrib.oed.glmm import known_covariance_linear_model
 
     # Note: coef is a p-vector, observation_sd is a scalar
     # Here, p=1 (one feature)
@@ -21,15 +23,15 @@ To create a classical Bayesian linear model, use::
 
 A non-linear link function may be introduced, for instance::
 
-    from pyro.contrib.glmm import logistic_regression_model
+    from pyro.contrib.oed.glmm import logistic_regression_model
 
     # No observation_sd is needed for logistic models
     model = logistic_regression_model(coef_mean=torch.tensor([0.]),
                                       coef_sd=torch.tensor([10.]))
 
 Random effects may be incorporated as regular Bayesian regression coefficients.
-For random effects with a shared covariance matrix, see :meth:`pyro.contrib.glmm.lmer_model`.
+For random effects with a shared covariance matrix, see :meth:`pyro.contrib.oed.glmm.lmer_model`.
 """
 
-from pyro.contrib.glmm.glmm import *  # noqa: F403,F401
-from pyro.contrib.glmm import guides  # noqa: F401
+from pyro.contrib.oed.glmm.glmm import *  # noqa: F403,F401
+from pyro.contrib.oed.glmm import guides  # noqa: F401
