+++ pyro/contrib/oed/util.py
@@ -2,7 +2,7 @@ import math
 import torch
 
 from pyro.contrib.util import get_indices
-from pyro.contrib.glmm import analytic_posterior_cov
+from pyro.contrib.oed.glmm import analytic_posterior_cov
 from pyro.infer.autoguide.utils import mean_field_entropy
