+++ tests/contrib/oed/test_glmm.py
@@ -5,7 +5,7 @@ from torch.distributions.transforms import AffineTransform, SigmoidTransform
 import pyro
 import pyro.distributions as dist
 import pyro.poutine as poutine
-from pyro.contrib.glmm import (
+from pyro.contrib.oed.glmm import (
     known_covariance_linear_model, group_linear_model, zero_mean_unit_obs_sd_lm,
     normal_inverse_gamma_linear_model, logistic_regression_model, sigmoid_model
 )
