+++ tests/contrib/oed/test_linear_models_eig.py
@@ -5,13 +5,13 @@ import pyro
 import pyro.distributions as dist
 import pyro.optim as optim
 from pyro.infer import Trace_ELBO
-from pyro.contrib.glmm import known_covariance_linear_model
+from pyro.contrib.oed.glmm import known_covariance_linear_model
 from pyro.contrib.oed.util import linear_model_ground_truth
 from pyro.contrib.oed.eig import (
     nmc_eig, posterior_eig, marginal_eig, marginal_likelihood_eig, vnmc_eig, laplace_eig, lfire_eig,
     donsker_varadhan_eig)
 from pyro.contrib.util import rmv, rvv
-from pyro.contrib.glmm.guides import LinearModelLaplaceGuide
+from pyro.contrib.oed.glmm.guides import LinearModelLaplaceGuide
 from tests.common import assert_equal
